// frontend/app/imaging/viewer/image-viewer/shared/types.ts
export interface ImageSize {
  width: number;
  height: number;
}
// ...
export interface Point {
  x: number;
  y: number;
}

/*
* 测量数据结构体, api/v1/measurements/{image_id} 和 api/v1/image-files/{image_id} 这两个接口都能用
* */
export interface MeasurementData {
  id: string;
  type: string;
  originalType?: string // api/v1/measurements/{image_id} 接口不提供此字段, api/v1/image-files/{image_id} 的 annotation 部分反序列化后的 measurements 字段提供此字段
  value: string;
  points: Point[];
  description?: string | null; // 这个 description 字段描述的是一个测量项做什么
  upperVertebra?: string | null; // api/v1/measurements/{image_id} 接口不提供此字段, 这里做兼容考虑
  lowerVertebra?: string | null; // api/v1/measurements/{image_id} 接口不提供此字段, 这里做兼容考虑
  apexVertebra?: string | null; // api/v1/measurements/{image_id} 接口不提供此字段, 这里做兼容考虑
}